+++ Changes.md
@@ -19,6 +19,14 @@ Breaking Changes
 
 - Gadget : Added new virtual method and private member variables.
 
+0.58.x.x (relative to 0.58.1.0)
+========
+
+API
+---
+
+- OpenImageIOReader : Added methods for controlling the maximum number of images files to keep open at one time.
+
 0.58.1.0 (relative to 0.58.0.1)
 ========